+++ docker-app/qfieldcloud/core/utils2/storage.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import logging
 import os
+import re
 from pathlib import PurePath
 from typing import IO, List, Set
 
@@ -148,6 +149,8 @@ def remove_user_avatar(user: "User") -> None:  # noqa: F821
     """
     bucket = qfieldcloud.core.utils.get_s3_bucket()
     key = user.useraccount.avatar_uri
+    if not key or not re.match(r"^users/\w+/avatar.(png|jpg)$", key):
+        raise RuntimeError("Suspicious S3 deletion")
     bucket.object_versions.filter(Prefix=key).delete()
 
 
@@ -200,6 +203,8 @@ def remove_project_thumbail(project: "Project") -> None:  # noqa: F821
     """
     bucket = qfieldcloud.core.utils.get_s3_bucket()
     key = project.thumbnail_uri
+    if not key or not re.match(r"^projects/\w+/meta/\w+.(png|jpg)$", key):
+        raise RuntimeError("Suspicious S3 deletion")
     bucket.object_versions.filter(Prefix=key).delete()
 
 
@@ -239,6 +244,11 @@ def purge_old_file_versions(project: "Project") -> None:  # noqa: F821
                 # but leaving it here as a security measure in case version
                 # ordering changes for some reason.
                 raise Exception("Trying to delete latest version")
+
+            if not old_version.key or not re.match(
+                r"^projects/\w+/.+$", old_version.key
+            ):
+                raise RuntimeError("Suspicious S3 deletion")
             # TODO: any way to batch those ? will probaby get slow on production
             old_version._data.delete()
             # TODO: audit ? take implementation from files_views.py:211
@@ -271,6 +281,8 @@ def upload_project_file(
 def delete_project_files(project_id: str) -> None:
     bucket = qfieldcloud.core.utils.get_s3_bucket()
     prefix = f"projects/{project_id}/"
+    if not not re.match(r"^projects/.+$", prefix):
+        raise RuntimeError("Suspicious S3 deletion")
     bucket.object_versions.filter(Prefix=prefix).delete()
 
 
@@ -344,6 +356,9 @@ def delete_file_version(
             versions_to_delete.append(file_version)
 
     for file_version in versions_to_delete:
+
+        if not re.match(r"^projects/\w+/.+$", file_version._data.object_key):
+            raise RuntimeError("Suspicious S3 deletion")
         file_version._data.delete()
 
         audit_suffix = "ALL" if is_deleting_all_versions else file_version.display
@@ -380,5 +395,6 @@ def get_stored_package_ids(project_id: str) -> Set[str]:
 def delete_stored_package(project_id: str, package_id: str) -> None:
     bucket = qfieldcloud.core.utils.get_s3_bucket()
     prefix = f"projects/{project_id}/packages/{package_id}/"
-
+    if not re.match(r"^projects/\w+/packages/\w+/$", prefix):
+        raise RuntimeError("Suspicious S3 deletion")
     bucket.object_versions.filter(Prefix=prefix).delete()
